refactoring mdr_execute_command_action -> mdr_process_speech_command_action
