+++ mdr_planning/mdr_actions/mdr_speech_actions/mdr_process_speech_command_action/ros/scripts/process_speech_command_client_test
@@ -11,25 +11,25 @@ import sys
 import rospy
 import actionlib
 
-from mdr_execute_command_action.msg import ExecutionAction, ExecutionGoal
+from mdr_process_speech_command_action.msg import ProcessSpeechAction, ProcessSpeechGoal
 
 
 def print_usage_info():
-    rospy.loginfo("usage: execute <command>")
+    rospy.loginfo("usage: process <command>")
     rospy.loginfo("       where <command> is a string")
 
 
 if __name__ == '__main__':
-    rospy.init_node('execute_client_test')
-    client = actionlib.SimpleActionClient('execute_command_server', ExecutionAction)
+    rospy.init_node('process_speech_client_test')
+    client = actionlib.SimpleActionClient('process_speech_command_server', ProcessSpeechAction)
     client.wait_for_server()
 
     try:
-        goal = ExecutionGoal()
+        goal = ProcessSpeechGoal()
         user_input_command = sys.argv[1]
         goal.command = user_input_command
         timeout = 15.0
-        rospy.loginfo("Sending action lib goal to execute_command_server, command: "
+        rospy.loginfo("Sending action lib goal to process_speech_command_server, command: "
                       + goal.command)
         client.send_goal(goal)
         client.wait_for_result(rospy.Duration.from_sec(int(timeout)))
