+++ mdr_planning/mdr_actions/mdr_speech_actions/mdr_process_speech_command_action/setup.py
@@ -4,7 +4,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
-   packages=['mdr_execute_command_action'],
-   package_dir={'mdr_execute_command_action': 'ros/src/mdr_execute_command_action'})
+   packages=['mdr_process_speech_command_action'],
+   package_dir={'mdr_process_speech_command_action': 'ros/src/mdr_process_speech_command_action'})
 
 setup(**d)
